+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Bugfixes
 
- - *tbd*
+ - The movie file searcher could crash if you have more than two movie directories (#1315, #1316)
 
 ### Changes